refactor: Refactored optim module
Separated optimizers from LR schedulers and optimizer wrappers.
