+++ holocron/optim/__init__.py
@@ -1,14 +1,11 @@
-from .lookahead import Lookahead
 from .lars import Lars
 from .lamb import Lamb
 from .radam import RAdam
 from .ralars import RaLars
-from .scout import Scout
+from . import wrapper
 from . import lr_scheduler
 
-del lookahead
 del lars
 del lamb
 del radam
 del ralars
-del scout
